Web App/requirements.txt
nltk==3.5
redis==3.5.3
newspaper3k==0.2.8
requests==2.24.0
google-cloud-language==2.0.0
Flask==1.1.2